Create Settings::LanguageUpdateService
Extends the base Settings::UpdateService in order to force the default
language on users whose current language setting wasn't available (or not
available anymore after the service's run).

Some spec factoring fit quite nicely here to extract the shared setup code
into a shared context.
