+++ spec/services/settings/language_update_service_spec.rb
@@ -1,0 +1,63 @@
+# frozen_string_literal: true
+
+# -- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) 2023 the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+# ++
+#
+
+require 'spec_helper'
+require_relative 'shared/shared_call_examples'
+require_relative 'shared/shared_setup_context'
+
+RSpec.describe Settings::LanguageUpdateService do
+  include_context 'with update service setup'
+
+  before do
+    # stub forcing default language
+    allow(instance)
+      .to receive(:force_user_language)
+  end
+
+  describe '#call' do
+    subject { instance.call(params) }
+
+    include_examples 'successful call'
+
+    it 'sets language of users having a non-available language to the default language' do
+      subject
+
+      expect(instance)
+        .to have_received(:force_user_language)
+    end
+
+    context 'when the contract is not successfully validated' do
+      let(:contract_success) { false }
+
+      include_examples 'unsuccessful call'
+    end
+  end
+end
